pull doc title section out into separate component
